Window Exit Bug
When choosing to record and then exiting, the tkinter window was persistant, it shows a blank window and the process is not proparly closed. However Screenshot feature works Perfectly as expected. so I used the same system exit code in the record function also and put the destroy function at the end. this seemed to have solved the issue. now after closing the recording by pressing q the tkinter window also closes as expected. Don't know weather the previous behaviour was intentional (so as to record/continue recording more) or not but i figured that it was unintentional as their was a empty tkinter root window still open. However this should solve it. btw Great Work Bro!!